+++ server/public/javascript/components/home/HomeComponent.js
@@ -1,4 +1,4 @@
-import { ce } from '../react-common.js';
+import { ce } from '../../react-common.js';
 import BinPreviewComponent from './BinPreviewComponent.js';
 
 /**
